instagram-notes-scraper/requirements.txt
instagrapi==2.1.1
python-dotenv==1.0.1
schedule==1.2.1
setuptools==69.5.1
